Add an API to coarsen/partition Targets by their cycles (#12251)
In order to recursively (such as for compilation) consume a graph of `Targets` containing cycles, we must "coarsen" the `Targets` in each cycle into a batch (called a [strongly connected component](https://en.wikipedia.org/wiki/Strongly_connected_component)).

This change adds a rule to compute the `CoarsenedTargets` for `Addresses`. To compile a graph of `Targets` with fine grained compiler invokes, you can "walk" the coarsened graph by requesting `CoarsenedTargets` for the roots, and then recursively requesting a compiler output per `CoarsenedTarget` dependency.