[Internal] Add `git_url()` helper to `docutil.py` (#12352)
For per-tool lockfiles in https://github.com/pantsbuild/pants/pull/12346, we will be embedding the default lockfile as a resource. We want to allow users to see what is in the lockfile, but don't want to dump all the contents to `./pants help`. 

Instead, this new `git_url()` util will allow us to link to the URL for the text file on GitHub.

[ci skip-rust]
[ci skip-build-wheels]