use `go list` to analyze stdlib imports (#12368)
## Problem

As described in https://github.com/pantsbuild/pants/issues/11910, the current version of the experimental Go plugin parses the directory structure of the Go distribution to determine the imports provided by the standard library. `go list` already provides this information in JSON form when asked. This was also how the v1 plugin obtained the same information. 

## Solution

Use `go list` to obtain import paths for the standard library. This PR introduces the base class `ResolvedImportPaths` which will be inherited by all types that represent import path mappings. The first use of this base class is for the `ResolvedImportPathsForGoLangDistribution` dataclass which represents analysis of the Go standard library.

## Testing

Added integration test for the import analysis and the first-party build integration test continues to pass.

Closes https://github.com/pantsbuild/pants/issues/11910.

[ci skip-rust]