Prefix the entire setup.py chroot. (#12372)
More specifically:

Add a prefix to the chroot. This prefix only exists
at build time, and does not affect the built dist,
but allows us to capture setuptools output from input
dirs without capturing the entire sandbox.

No longer prefix just the sources within the chroot.
This was never particularly necessary, and I originally
implemented that somewhat arbitrarily. Note that this means
that from now on dists will have a different internal structure,
but one that is more usual and simpler, so probably better.

Run setup.py with its directory as the cwd. Custom setup.pys
will expect that.

This will be useful when we run the setuptools develop command
on setup.py scripts that build native code and write the .so or
other output files into the input dirs.

This is a replay of #12359 with a fix for the issue that caused it
to be reverted.

[ci skip-rust]