Dependency inference for `pex_binary` and `python_awslambda` can disambiguate ambiguous modules based on file paths (#12326)
A user has the files `root1/app.py` and `root2/app.py`. In both folders, they have a `pex_binary` with `entry_point="app.py"`. Our dep inference code complains this is ambiguous because >1 target exports the module `app`.

But we can leverage this requirement for the `pex_binary` and `python_awslambda` targets to do better than that:

> Like with the sources field, file paths are relative to the BUILD file, rather than the build root.

This assumption means that we know—when using a file name—that file must be in the current dir of the target or a subdir. That implies that all owning targets of the file must be in the current dir or an ancestor dir. We can safely ignore all candidate targets out of that tree.

[ci skip-rust]
[ci skip-build-wheels]