[internal] Don't build fs_util in PR builds (#11897)
We don't release the built artifact in PR builds, so we were only doing this to smoke test our release process. However, the fs_util release is pretty simple and also fairly low stakes, unlike our normal pantsbuild/pants release. Rust compilation already catches most issues. We will catch if the release is broken thanks to branch builds. 

It's not worth the cost of resources and time to run this smoke test in PR builds.