introduce MaybeEmptySiblingAddresses address spec (#12407)
Introduce the `MaybeEmptySiblingAddresses` address spec to allow searching for addresses in a directory without a BUILD file or for which there are no targets in any of the BUILD files. It is similar in spirit to the already-existing `MaybeEmptyDescendantAddresses`.

This PR is necessary for the Go tailor support in https://github.com/pantsbuild/pants/pull/12406. The inference for `go_external_module` attempts to find `go_module` targets along the tailor `search_paths`, but cannot use the `SiblingAddresses` address spec because directories at the top of the tree that may not have a BUILD file will cause `SiblingAddresses` to raise an exception.

`SiblingAddresses` will now fail if it does not find any matching targets.

[ci skip-rust]

[ci skip-build-wheels]