Add special casing for '--javac-jdk=system' as a temporary hack for #12293.
This causes JavacBinary to pass the `--system-jvm` option to Coursier,
which selects whichever JDK Coursier happens to find already installed
on the system.  Long term, we shouldn't use this behavior in tests and
probably shouldn't even allow it at all since the inherent non-hermeticity
isn't properly captured in cache keys, but for now it allows testing
of Coursier-dependent backends in CI until #12293 is resolved properly.