tailor: work-around targets that do not have a `sources` field (#12409)
## Problem

`./pants tailor` currently assumes that all targets on which it operates will have a `sources` field. As part of the Go plugin's tailor support in https://github.com/pantsbuild/pants/pull/12406, I would like `tailor` to infer `go_external_module` targets which do not have a `sources` field from the dependencies of `go_module` targets.

Attempting to add a `GoExternalModule` target as a `PutativeTarget` caused invocation of `tailor` to raise an exception in `default_sources_for_target_type` because that function could not find a `sources` field.

## Solution 

Work-around the issue by teaching `tailor` to ignore the lack of a `sources` field if there are no owned sources or triggering sources for the `PutativeTarget`. The exception is now raised in `PutativeTarget.for_target_type` if there are any such sources. `default_sources_for_target_type` just returns an empty tuple if it could not find a `sources` field.

This is just a work-around and is not intended as a longer term solution for `tailor` to support target types without `sources`.

[ci skip-rust]

[ci skip-build-wheels]