[internal] Port `PyDigest` and `PySnapshot` to PyO3 (#12422)
We are not yet ready to use the PyO3 implementation in production because this type gets used with the Scheduler code, e.g. our intrinsic rules. 

Instead, this implements the types in PyO3 while still keeping rust-CPython. It adds unit tests to assert the functional equivalence of both types. 

Even though this PR does violate DRY, I argue that it's worth it to help make the PyO3 migration more incremental and less risky.

[ci skip-build-wheels]