[internal] Add generate_all_lockfiles.py to correctly generate default tool lockfiles (#12418)
The requirements we use for pantsbuild/pants might be different than what makes sense for the default we provide to our users. For example, we currently use two Flake8 plugins that should not be in the default lockfile.

This script allows us to be confident that the default lockfiles we generate are entirely decoupled from our own internal settings.