Shield isolated-Sessions / StreamingWorkunitHandlers from Ctrl+C. (#12483)
Isolated `Session`s are used for `StreamingWorkunitHandler` callbacks so that they are able to be cancelled independently of the foreground work that they are monitoring from a background thread. But when run without `pantsd`, _all_ `Session`s (including those for background work) were being cancelled when SIGINT was received. This was causing `SWH` callbacks which needed to interact with the `Scheduler` to receive `KeyboardInterrupt`s while they were tearing down after a run had been cancelled.

To fix this, we shield isolated `Session`s from being cancelled by Ctrl+C, and expand `SWH` testing to cover it.

[ci skip-build-wheels]