Go: refactor Go SDK invocation into common rule (#12494)
Introduce `GoSdkProcess` type and a rule to invoke the Go SDK using it. Then refactor invocations of the Go SDK to use this common type. This removes the duplication of the setup code needed to invoke the Go SDK.

Caveat: The `import_analysis` module is not ported because it wants to know the digest of the Go SDK. Will port that module in a subsequent PR.