Resolve plugins using the PEX --python option. (#12500)
As described in #12361, when Pants is running under a Python (generally selected by the `pants` script) which is not part of the `interpreter_search_path`, plugin resolution will fail to find an exact match.

Resolve Pants plugins using the PEX `--python` option to ensure that they are located regardless of the `--python-path`.

Fixes #12361 and the repro described there.

[ci skip-rust]
[ci skip-build-wheels]