[internal] Lockfiles: automatically say how to regenerate them header (#12501)
The semantics are still not clear for how users should regenerate lockfiles. But in the meantime, this gives slightly more insight than before and gets a code skeleton into place.

This adds an option to allow users to override the regeneration command, which we use for pantsbuild/pants. The idea comes from pip-compile having a similar option.

*Note that the command to regenerate the user lockfile is usually going to be incorrect, as we're not assuming they ran `./pants lock ::`, which is not always the case. Multiple user lockfiles is the next part of the project, and this is all feature-gated, so I didn't want to spend too much time trying to get this regen command to be accurate.

[ci skip-rust]