Change the PythonProtobufMypyPlugin option scope. (#12525)
top-level scopes are not allowed to contain dots. This is only weakly enforced however,
when calling get_scope_info() on the Subsystem instance. So this snuck by.

But it will fail the validation check if, for example, calling ./pants help-all with the
python proto backend enabled.

In any case, the name was over-elaborate, and this new name is more succinct,
more intuitive and more conventional (i.e., it is named after the tool it configures).

[ci skip-rust]

[ci skip-build-wheels]