[internal] Add `InterpreterConstraints.flatten()` in preparation for Poetry lockfiles (#12543)
## Problem

We allow ICs to be a list, where each element is ORed. For example, `['==2.7.*', '==3.6.*']` means "2.7 OR 3.6". 

However, Poetry expects the interpreter constraints to be a single constraint, and it has no notion of "OR". So, we need an automated way to convert into a format understood by Poetry.

## Algorithm strategy

Find the upper and lower bounds of the interpreter constraints, while also finding any versions to skip. With this, we can Include the entire possible range, but blocklist any skipped versions.

For example, `['==3.3.*', '==3.5.*.']` becomes `>=3.3,<3.6,!=3.4.*'`.

Where possible, we try to make the result more readable, like preferring `!=3.7.*` over `!=3.7.0,!=3.7.1,...`.