Fix shlexing of passthrough args. (cherrypick of #12547) (#12550)
`./pants run example.py -- -flag "some args"` currently results in a different `sys.argv` for the process than `./example.py -flag "some args"` would. This was because we were applying shlexing after the source/rank of an option had been discarded, which was too late to determine whether to shlex (flag/config/env values should be shlexed: passthrough arguments should not).

Move `shell_str` `list` shlexing into `ListValueComponent` such that it is skipped when a `ListValueComponent` is manually constructed for passthrough args.

One downside of this change is that list-value addition and subtraction will now apply to the shlexed tokens of a string, rather than to the entire string (i.e `+["some args"],-["args"]` will result in a final set of `["some"]`). But I expect that list subtraction is rare for `shell_str` args.

[ci skip-rust]
[ci skip-build-wheels]