[internal] Refactor MyPy config file setup (#12593)
`mypy/rules.py` is really complex and I'm having a hard time understanding it while adding a tool lockfile. This is the first of a couple refactors.

Here, we better encapsulate the config file code by moving into `mypy/subsystem.py` with a dedicated rule. This allows us to make the unit test more comprehensive.

[ci skip-rust]
[ci skip-build-wheels]