Embrace that help strings are markdown. (#12606)
Previously we sometimes wanted help strings treated as markdown and sometimes as plain text,
depending on author convenience. Or we wanted parts of a given string treated in different ways.
We tried various hacks to anticipate author intention, but they never worked properly, and caused
no end of errors and edge cases and gotchas.

Now we simply accept that help strings are treated as markdown. This makes sense: We need
them to be in a format that is human-readable (for when displaying help on the CLI, or when reading
source code), and also machine-readable, for rendering by readme.com. Markdown is such a
dual format.

This change:

Removes the weird heuristics.
Fixes up various help strings to conform with the new reality.
The big gotcha now is that authors of help strings must be aware of this fact. For example, you will
often want to surround things that look like markup or HTML tags, but are not, with backticks.

I have manually eyeballed a large subset of the docs generated with these changes, including
the "known offenders", and this seems like a good way to go.

A possible future change is to detect when help strings have changed, render them, and prompt
the author to manually verify that the rendered pages look good.

[ci skip-rust]

[ci skip-build-wheels]