Remove the `rationale` field from BashBinaryRequest.
It exists only to provide slightly more context for an error
message. But this context isn't actionable (the rationale doesn't
affect how you fix the error), and the downside is that the
rationale affects the cache key, so we have to re-find bash
for every rationale, which is superfluous extra work.

If in the future someone, naively, makes the rationale a dynamic
message, then we will thrash for unclear reasons. Better to remove
this landmine preemptively.

[ci skip-rust]

[ci skip-build-wheels]