Stop using `[python-setup].requirements_constraints` when installing tools like Black and MyPy (#12639)
It was an oversight (by me!) that `--requirement-constraints` was being used for tool installations. You may want for `black` to use a different version of `ast` than your own code is using, for example. It was bad coupling, and not documented or transparent to the user.

Instead, we have the much better feature of tool lockfiles: https://github.com/pantsbuild/pants/issues/11898, which will soon be promoted to be stable in Pants 2.7. If you want to lock down a tool, you need to use this new feature rather than the constraints file.

This is technically a breaking change, but the past was so broken that we consciously do not use a deprecation.

[ci skip-rust]
[ci skip-build-wheels]