tailor: do not append comma to kwarg with empty tuple (#12653)
If an empty tuple is passed as the value of a kwarg to `PutativeTarget.for_target_type`, `tailor` will generate `[,]` in the `BUILD` file which does not parse. Fix the issue by only appending the comma when there are actually values to put into the BUILD file for the kwarg.