[internal] Test that Python tools work with all expected interpreter versions (#12656)
First part of https://github.com/pantsbuild/pants/issues/12601: we need to test that the default lockfile works for all expected Python versions. It's important that we not only install the tool PEX, but also run it to make sure there are no missing runtime dependencies.

_These tests do not yet actually consume the lockfile_, but that will happen soon once we change the default to be `<default>`. This is prework.

These tests will fail if the expected Python interpreters are not available in the environment. We definitely want to error in CI in that case, so we can fix CI to have the relevant interpreters. But this does mean `./pants test ::` is less likely to work for desktop use.

This does not test iPython and Setuptools because they don't have tests set up already: https://github.com/pantsbuild/pants/issues/12658 and https://github.com/pantsbuild/pants/issues/12657.

[ci skip-rust]
[ci skip-build-wheels]