[internal] Rename `./pants lock` to `./pants generate-user-lockfile` and `./pants tool-lock` to `./pants generate-lockfiles` (#12641)
We currently have separate goals for user lockfiles vs tool lockfiles because the implementations are different right now, e.g. specs being different. It's easier to keep separate while iterating. We will unify them, but are not yet ready.

We want to release tool lockfiles as stable in Pants 2.7, so it should be using the final proper name of `./pants generate-lockfiles` rather than `./pants tool-lock`.

Once user lockfiles are stable in 2.8, we will kill `generate-user-lockfile` in favor of a unified `generate-lockfiles` goal.

## `generate-lockfiles` vs. `lock` vs `resolve` vs. something else

See https://github.com/pantsbuild/pants/pull/12641#issuecomment-905777885 for why I went with `generate-lockfiles`, and the above thread for alternative proposals.

[ci skip-rust]
[ci skip-build-wheels]