Add `--generate-lockfiles-resolve` to allow generating for only certain tools (#12676)
Closes https://github.com/pantsbuild/pants/issues/12640.

## Named resolves

This introduces the notion of "named resolves": associating a particular lockfile with a name. For tools, the name is the options scope like `black` and `mypy-protobuf`. For multiple user lockfiles, the names will be user defined like `py2` and `data-science`. Users will reference lockfiles by resolve name, rather than lockfile path.

Here, users can now run `./pants generate-lockfiles --resolve=black --resolve=isort` for example. In the future, they can run `./pants generate-lockfiles --resolve=black --resolve=py2 --resolve=data-science`. If the option isn't specified, we generate for all resolves.

## Unrecognized named resolves

It's unfortunate that we cannot dynamically list all valid named resolves in `./pants help`, as the options system can't execute the rules API. Instead, we have a good error message when the resolve isn't recognized:

> pants.backend.python.goals.lockfile.UnrecognizedResolveNamesError: Unrecognized resolve name from the option `--generate-lockfiles-resolve`: fake
>
> All valid resolve names: ['black', 'coverage-py', 'docformatter', 'flake8', 'ipython', 'isort', 'mypy', 'pytest', 'setuptools']

## Warning about lockfiles not enabled

If a user tries to generate a lockfile for a tool that is either disabled w/ `<none>` or the default with `<default>`, we'll warn:

> You requested to generate a lockfile for isort because you included it in `--generate-lockfile-resolve`, but `[isort].lockfile` is set to `<default>` so a lockfile will not be generated.
>
> If you would like to generate a lockfile for isort, please set [isort].lockfile` to the path where it should be generated and run again.