support indivisible Sources fields (#12685)
## Motivation

As part of #12643 (which adds support for building third-party Go code and linking into a binary), the Go plugin needs to obtain only unexpanded targets since file-level subtargets have no meaning semantically for Go code since the "unit of build" is an entire package, not an individual file.

The Target API currently eagerly expands build targets into file-level subtargets. This behavior requires work-arounds in the Go plugin in order to recover the original build target on which the Go plugin operates.

## Solution

Introduce the notion of "indivisible" Sources fields. When a `Sources` field is marked `indivisible = True`, then the Target API will not expand that target into file-level subtargets. This marker should allow the Go plugin to not require work-arounds to recover build targets because it won't ever be given per-file subtargets.

This change also benefits other plugins where the "unit of build" is an entire target and not files; for example, the experimental Terraform plugin where Terraform operates on entire directories at once.

See the discussion in https://github.com/pantsbuild/pants/pull/12677 for more context.

[ci skip-rust]

[ci skip-build-wheels]