Decompose requirements pexes (#12675)
As described in #12548 (and further clarified in https://github.com/pantsbuild/pants/issues/12548#issuecomment-906906106), when we extract subsets of a `repository.pex` (which is itself built from a lockfile or constraints file), each of the subsets is likely to be highly redundant. And because the subsets cannot be structure/content shared by the CAS, the result is an increase of storage usage proportional to the number of distinct subsets. Concretely: in some repositories, we observed a ~45X blowup of storage due to ~90 different subsets being constructed.

To fix this, this change moves to constructing each `requirements.pex` subset of a `repository.pex` from a PEX_PATH containing one PEX per root requirement. As implemented here, the storage blowup is reduced to ~4.4X, but as mentioned in #12688, these "single-requirement" PEXes are currently transitive, and it's possible that this could be fixed in the future. 

An alternative to this change would have been to remove the "subsetting" step entirely, and to directly use the entire `repository.pex` in consumers (even if they only used a small fraction of it). But that would have the disadvantage that any change to any requirement would invalidate all consumers in a repository, regardless of whether the requirement was relevant to them, and would additionally violate hermeticity by allowing, for example, `test`s to observe resources that they don't have dependencies on. 

Fixes #12548.

[ci skip-rust]