Use Pants's default tool lockfiles by default for Python tools like Black, Pytest, and Flake8 (#12689)
This is a breaking change: if you have changed the tool's `version` or `extra_requirements`, by default Pants will now error telling you to regenerate a lockfile (https://github.com/pantsbuild/pants/issues/12654). We are consciously breaking the Deprecation Policy because we believe the past of no lockfile (and using `[python-setup].requirement_constraints`) was very broken, and that it's crucial we have a sensible default. We'll try to document this breaking change well. 

[ci skip-rust]
[ci skip-build-wheels]