[internal] Remove unregistered `get_buildroot` and `buildfile_path` symbols from BUILD files (#12746)
We accidentally weren't registering these. That's just fine: there shouldn't be a need for absolute paths in BUILD files because v2 uses relative paths for everything (REAPI compliance iiuc). 

If a user has a need for the BUILD file path, they should use the Rules API to do that. We want to stop using `objects`.

[ci skip-rust]
[ci skip-build-wheels]