Fix UI rendering when a workunit has completed children but no running children (#12748)
As observed in #12732: when a running workunit had completed children but no running children, it would not be rendered, since it was not a leaf of the workunit graph.

It's possible that we never noticed this before because we tend to create "complete" trees of workunits: i.e. that most of the time below one workunit is taken up by other workunits, rather than having a child cover only a portion of parent's time (as in #12721). But this couldn't always have been the case: after a parent's children complete, it will always run for some amount of time before completing itself: this likely caused flickering in the UI.

This change filters the graph that we maintain for UI purposes to contain only running (Started but not Completed) workunits. In order to remove workunits from the graph when they complete without affecting node ids, we switch to using `StableGraph`. With that bug fixed, also restores the workunit added by #12721.

[ci skip-build-wheels]