add intrinsic to produce a `DigestEntries` from a `Digest` which reference file content by hash (#12693)
## Motivation

The Go and JVM plugins both have a need to create a `Digest` dynamically from files from another `Digest`, but without needing to actually manipulate the content bytes of the files in the `Digest`. The engine's Filesystem API, however, only provides the `FileContent` and `DigestContents` types for obtaining information about file content, and the `CreateDigest` intrinsic only accepts `FileContent` and `Directory` instances. This means that both plugins are asking the engine to materialize file content into memory that will not ever be inspected by Python code. (Indeed, this is what the plugins are currently doing as a work around despite the performance problems.)

## Solution

Introduce the `FileEntry` type which is similar to `FileContent` but refers to the file's content using a `FileDigest` instead of including the file's actual content bytes. The PR also adds a `DigestEntries` type which can be requested via a new intrinsic to obtain the collection of `Union[FileEntry, Directory]` objects for a `Digest`. (Files are represented by `FileEntry`, and empty leaf directories by `Directory` in a similar manner to `CreateDigest`.)

The `CreateDigest` intrinsic is updated to accept `FileEntry` instances along with already-accepted `FileContent` and `Directory` instances.

[ci skip-build-wheels]