A proper access class for the readme.com API.
We're currently only using the API to generate the reference docs,
and were doing so somewhat haphazardly.

But there is an increasing need for tools to sync updates across
versions, especially as we're about to make some changes
to the getting started docs. These will be easier to create with
a proper abstraction.

Note that this class only provides the parts of the API we care
about.

[ci skip-rust]

[ci skip-build-wheels]