[internal] Simplify PEX setup for `./pants run` (#12806)
To implement https://github.com/pantsbuild/pants/issues/11619, we need to set `--console-script` on the requirements PEX, not the runner PEX. It does not work to set on the runner PEX because it will not have any distributions available. 

In trying to fix that, I realized it is useful to simplify no longer having a requirements PEX vs. runner PEX. Originally, we split this so that the requirements PEX shared the cache between goals like `test` and `lint`. That is less relevant now that we first resolve from the constraints/lockfile, then extract the subset of deps - most of the heavy lifting will already be shared.

*https://github.com/pantsbuild/pants/pull/12806 was necessary to land before for this to work. Otherwise, the lockfile.pex would have much less reuse because of setting `additional_args` on the `PexFromTargetsRequest`.

[ci skip-rust]
[ci skip-build-wheels]