Upgrade to poetry 1.1.8 (#12856)
https://python-poetry.org/history/#118---2021-08-19

Fixed an error with repository prioritization when specifying secondary repositories. (#4241)
Fixed the detection of the system environment when the setting virtualenvs.create is deactivated. (#4330, #4407)
Fixed the evaluation of relative path dependencies. (#4246)
Fixed environment detection for Python 3.10 environments. (#4387)
Fixed an error in the evaluation of in/not in markers (python-poetry/poetry-core#189)