Rename the `typecheck` goal to `check` (#12858)
Compiled languages need a goal that semantically means: "the code compiles". 

Although #11249 originally proposed exposing a `build` goal to generically expose compiled outputs, there was consensus that there is value in a goal with no outputs other than indicating whether the code successfully compiled. Unlike with a `test` or `package` goal, where the output is test results or an executable (respectively), the only expected output for this goal is the compiler or typechecker output.

As discussed in #11249, Rust's [`check` goal](https://blog.rust-lang.org/2017/03/16/Rust-1.16.html#whats-in-1160-stable) has a related semantic meaning: it runs the minimum amount of compilation necessary to validate the code, _without_ producing an executable. Because it completes ~10x faster than full compilation while catching >99% of errors, it's an excellent default mode to iterate in. It also has a lot in common with pure/non-compiler typecheckers like `mypy`, which usually don't qualify as "linters" due to their cost.

To align with the nascent convention, and to prepare to expose "check mode" compiles for other languages like `go` and `java`, this change renames the `typecheck` goal to `check`.

Fixes #11249.