Fix setup.py path in test.
The path is supposed to be relative to the BUILD file, but was
specified relative to the buildroot.

This test passed because src/python/foo/src/python/foo/setup.py
becomes foo/setup.py after stripping source roots (because
src/python/foo/src/python/ is recognized as a source root).
But that was still a bad example for any readers.

[ci skip-rust]

[ci skip-build-wheels]