[internal] java: make junit output parsing in tests more robust (#12909)
When I was running the JUnit integration test, the test runner output was different than what the junit integration tests expected because JUnit saw a different terminal type on macOS than in CI/Linux and rendered the `[OK]` marker as a Unicode checkmark. The integration tests were hard-coded to look for the `[OK]` which caused them to fail.

For now, force the JUnit runner to use ASCII output and update the integration test accordingly. Add comments to note that the work-around should be removed when https://github.com/pantsbuild/pants/issues/12812 implements parsing the junit.xml.

This PR was split as pre-work from #12897.

[ci skip-rust]

[ci skip-build-wheels]